* code refactoring for more easier further development
