+++ filer/admin/common_admin.py
@@ -4,10 +4,7 @@ from django.contrib.admin.util import unquote
 from django.core.urlresolvers import reverse, resolve
 from django.http import HttpResponseRedirect
 from filer.models import Folder
-from filer.admin.tools import (has_admin_role, has_role_on_site,
-                               has_admin_role_on_site,
-                               get_admin_sites_for_user,
-                               get_sites_for_user,)
+from filer.admin.tools import (has_admin_role, has_role_on_site)
 from filer.views import (popup_param, selectfolder_param, popup_status,
                          selectfolder_status, current_site_param)
 
@@ -113,19 +110,7 @@ class FolderPermissionModelAdmin(CommonModelAdmin):
                 return True
         else:
             folder = Folder.objects.get(id=folder_id)
-            # nobody can add subfolders in core folders
-            if folder.is_readonly():
-                return False
-            # only site admins can add subfolders in site folders with no site
-            if not folder.site and has_admin_role(request.user):
-                return True
-            # regular users need to have permissions to add folders and
-            #   need to have a role over the site owner of the folder
-            can_add = super(FolderPermissionModelAdmin, self).\
-                has_add_permission(request)
-            if (folder.site and can_add
-                    and has_role_on_site(request.user, folder.site)):
-                return True
+            return folder.has_add_permission(request.user)
         return False
 
     def has_change_permission(self, request, obj=None):
@@ -136,20 +121,7 @@ class FolderPermissionModelAdmin(CommonModelAdmin):
         if not folder:
             return request.user.has_perm('filer.can_use_directory_listing')
 
-        # nobody can change core folder
-        if folder.is_readonly():
-            return False
-        # only admins can change site folders with no site owner
-        if not folder.site and has_admin_role(request.user):
-            return True
-
-        if folder.site:
-            if not folder.parent:
-                # only site admins can change root site folders
-                return has_admin_role_on_site(request.user, folder.site)
-            return can_change and has_role_on_site(request.user, folder.site)
-
-        return False
+        return folder.has_change_permission(request.user)
 
     def has_delete_permission(self, request, obj=None):
         folder = obj
@@ -159,20 +131,7 @@ class FolderPermissionModelAdmin(CommonModelAdmin):
         if not can_delete or not folder:
             return can_delete
 
-        if folder.is_readonly():
-            return False
-
-        # only admins can delete site folders with no site owner
-        if not folder.site and has_admin_role(request.user):
-            return True
-
-        if folder.site:
-            if not folder.parent:
-                # only site admins can change root site folders
-                return has_admin_role_on_site(request.user, folder.site)
-            return can_delete and has_role_on_site(request.user, folder.site)
-
-        return False
+        return folder.has_delete_permission(request.user)
 
     def can_view_folder_content(self, request, folder):
         if folder.is_readonly():
@@ -186,6 +145,7 @@ class FolderPermissionModelAdmin(CommonModelAdmin):
 
         return False
 
+
 class FilePermissionModelAdmin(CommonModelAdmin):
 
     def _get_parent_for_view(self, obj):
@@ -196,41 +156,11 @@ class FilePermissionModelAdmin(CommonModelAdmin):
             has_change_permission(request, obj)
         if not can_change or not obj:
             return can_change
-
-        folder = obj.folder
-        if not folder:
-            # clipboard and unfiled files
-            return True
-
-        if folder.is_readonly():
-            # nobody can change core folder
-            # leaving these on True based on the fact that core folders are
-            # displayed as readonly fields
-            return True
-        # only admins can change site folders with no site owner
-        if not folder.site and has_admin_role(request.user):
-            return True
-
-        if folder.site:
-            return has_role_on_site(request.user, folder.site)
-
-        return False
+        return obj.has_change_permission(request.user)
 
     def has_delete_permission(self, request, obj=None):
         can_delete = super(FilePermissionModelAdmin, self).\
             has_delete_permission(request, obj)
         if not can_delete or not obj:
             return can_delete
-        folder = obj.folder
-        if not folder:
-            # clipboard and unfiled files
-            return True
-        # nobody can delete core files
-        if folder.is_readonly():
-            return False
-        # only admins can delete site files with no site owner
-        if not folder.site and has_admin_role(request.user):
-            return True
-        if folder.site:
-            return has_role_on_site(request.user, folder.site)
-        return False
+        return obj.has_delete_permission(request.user)
